Fixed a few issues

+++ src/engine/RenderList.cpp
@@ -176,7 +176,7 @@ namespace Atlas {
         if (!currentMatricesBuffer || currentMatricesBuffer->size < sizeof(mat4) * currentActorMatrices.size()) {
             auto newSize = currentMatricesBuffer != nullptr ? currentMatricesBuffer->size * 2 :
                            sizeof(mat4) * currentActorMatrices.size();
-            newSize = std::max(newSize, 1ul);
+            newSize = std::max(newSize, size_t(1));
             auto bufferDesc = Graphics::BufferDesc {
                 .usageFlags = VK_BUFFER_USAGE_STORAGE_BUFFER_BIT,
                 .domain = Graphics::BufferDomain::Host,
@@ -189,7 +189,7 @@ namespace Atlas {
         if (!impostorMatricesBuffer || impostorMatricesBuffer->size < sizeof(mat4) * impostorMatrices.size()) {
             auto newSize = impostorMatricesBuffer != nullptr ? impostorMatricesBuffer->size * 2 :
                            sizeof(mat4) * impostorMatrices.size();
-            newSize = std::max(newSize, 1ul);
+            newSize = std::max(newSize, size_t(1));
             auto bufferDesc = Graphics::BufferDesc {
                 .usageFlags = VK_BUFFER_USAGE_STORAGE_BUFFER_BIT,
                 .domain = Graphics::BufferDomain::Host,
